Implement exercise 1.3.37
Queue-based solution to Josephus problem.

In the Josephus problem from antiquity, N people are in dire straits and
agree to the following strategy to reduce the population.

They arrange themselves in a circle (at positions numbered from 0 to
N-1) and proceed around the circle, eliminating every Mth person until
only one person is left. Legend has it that Josephus figured out where
to sit to avoid being eliminated.

This is a Queue-based implementation which calculates the sequence in
which people are elimated, and thus can be used to show Josephus where
to sit in the circle.